Move sorter function to storage service Ticket: #33558
